requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
flask-caching==2.0.2
Pillow==10.0.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-image==0.21.0
onnxruntime==1.15.1